refactor: move fetchAPI file to service folder
